feat(booking-management): implement booking room state management and UI components
- Added BookingRoomState and its subclasses for managing booking states.
- Created BookingConfirmationAction widget for handling booking confirmation actions.
- Developed BookingConfirmationBody widget to display booking details and actions.
- Implemented BookingConfirmationDialog for presenting booking confirmation UI.
- Added PrintFatoraButton for printing invoices related to bookings.
- Introduced ExtendTheStayCubit and ExtendTheStayState for managing stay extensions.
- Created RoomsFilterHelper for filtering room entities based on various criteria.
- Developed RoomsFormController for managing room form inputs and states.
- Implemented custom dropdown button field for selecting options in forms.
- Created ExtendTheStayDialog and its body for handling stay extension UI.
- Developed FloorSection widget for displaying and selecting room floors.

+++ lib/core/services/local_database.dart
@@ -1,3 +1,4 @@
+import 'dart:developer';
 import 'dart:io';
 import 'package:path/path.dart' as p;
 import 'package:flutter/foundation.dart';
@@ -26,7 +27,14 @@ class DatabaseHelper {
     final dbPath = await getDatabasesPath();
     final path = p.join(dbPath, fileName);
 
-    return await openDatabase(path, version: 1, onCreate: _createDB);
+    return await openDatabase(
+      path,
+      version: 1,
+      onCreate: _createDB,
+      onConfigure: (Database db) async {
+        await db.execute('PRAGMA foreign_keys = ON');
+      },
+    );
   }
 
   Future _createDB(Database db, int version) async {
@@ -56,6 +64,7 @@ class DatabaseHelper {
       paidAmount REAL NOT NULL DEFAULT 0,
       paidType TEXT NOT NULL,
       notes TEXT,
+      stutasBooking TEXT,
       FOREIGN KEY (roomID) REFERENCES rooms (roomId) ON DELETE CASCADE
     )
   ''');
@@ -92,8 +101,8 @@ class DatabaseHelper {
   Future<int> updateData({
     required String table,
     required Map<String, dynamic> row,
-    required String idColumn,
-    required dynamic id,
+    String? idColumn,
+    dynamic id,
   }) async {
     final db = await instance.database;
     return await db.update(table, row, where: '$idColumn = ?', whereArgs: [id]);
@@ -102,11 +111,16 @@ class DatabaseHelper {
   // 🟢 Delete
   Future<int> deleteData({
     required String table,
-    required String idColumn,
-    required dynamic id,
+    String? idColumn,
+    dynamic id,
   }) async {
     final db = await instance.database;
-    return await db.delete(table, where: '$idColumn = ?', whereArgs: [id]);
+    if (idColumn != null) {
+      log("🟢 Deleted from '$table' where $idColumn = $id");
+      return await db.delete(table, where: '$idColumn = ?', whereArgs: [id]);
+    }
+    log("🟢 Deleted database '$table'");
+    return await db.delete(table);
   }
 
   // 🟢 Close DB
